refactor: path to import controllers
