+++ backend/v1/controllers/index.ts
@@ -1,0 +1,5 @@
+import { AirPortController } from './airport'
+import { DestinationController } from './destination'
+import { ReservationController } from './reservation'
+
+export { AirPortController, DestinationController, ReservationController }
